Title: 달에서 무게 구하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240306/달에서 무게 구하기/weight-on-the-moon.cpp
@@ -5,7 +5,7 @@ int main() {
 
 	
 
-	double a = 13;
+	int a = 13;
     double b = 0.165;
     double c = a*b;
